+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     emails.
   - Tests refactor to make them more specific to each type of data, mocking interfaces between different classes.
 - #59 - Added a new parser `EmailDateParser` that uses the temail `Date` to get the `Stamp` and use in most of the `Providers` via the `CombinedProcessor`. Also, `Maintenance.stamp` attribute is mandatory.
+- #60 - Added new provider `Seaborn` using `Html` and a new parser for Email Subject: `EmailSubjectParser`
 
 ### Fixed